Gameplay Blueprints Created
The Mission BPs are created and the drop off collision boxes aswell. They can be altered and the material outline hasnt been added yet. still waiting on the proper meshes. Player can now hold click to lower and elevate the orb. UI place holder for the checklists added for the designers to alter.